pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datenerfassung"
version = "0.1.0"
description = "Self-hosted ingest pipeline for household bookkeeping (receipts, OCR, normalization, rule-based categorization)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "pydantic>=2.7",
  "pyyaml>=6.0.1",
  "python-multipart>=0.0.12",
  "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
  "mypy>=1.10",
]
ocr = [
  "paddleocr>=2.8.0",
  "paddlepaddle; python_version < '3.13'",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
